Added LSTM IMDb review classifier
This commit comprises the addition of three files for an IMDb review classifier. The 'lstmIMDb.py' script is responsible for training an LSTM model using the IMDb dataset of positive and negative movie reviews. The script utilizes PyTorch and TorchText and, if possible, scans for GPU availability to accelerate the training process. The 'lstmIMDbtest.py' script loads the trained model and defines a prediction function for a given review's sentiment. In addition, an example usage of the function is provided. The 'textclass.pth' file contains the trained model in PyTorch's native format. Together, these files enable users to train their own sentiment classifier and estimate the sentiment of movie reviews.